requirements.txt
# Core Flask dependencies
flask>=3.0.0
flask-sqlalchemy>=3.1.1
flask-login>=0.6.3
flask-mail>=0.9.1
flask-migrate>=4.0.5
flask-wtf>=1.2.1

# Security and password hashing
werkzeug>=3.0.1

# HTTP requests for TMDB API
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Data processing and machine learning (for future ML features)
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0

# Database migration tool
alembic>=1.13.0

# Testing framework
pytest>=8.0.0

# Email validation (optional but recommended)
email-validator>=2.1.0

# For better datetime handling
python-dateutil>=2.8.2

# PostgreSQL adapter for Supabase
psycopg2-binary>=2.9.9

# Production server
gunicorn>=21.2.0 